Create PROJECT-BLACKJACK
Blackjack, formerly also Black Jack and Vingt-Un, is the American member of a global family of banking games known as Twenty-One, whose relatives include the British game of Pontoon and the European game, Vingt-et-Un.[1] It is a comparing card game between one or more players and a dealer, where each player in turn competes against the dealer. Players do not compete against each other. It is played with one or more decks of 52 cards, and is the most widely played casino banking game in the world.[2]:342
https://en.wikipedia.org/wiki/Blackjack